Use TokenRange

+++ Api/src/UrlShortener.Code/Base62EncodingExtensions.cs
@@ -4,7 +4,7 @@ public static class Base62EncodingExtensions
 {
     private const string Alphabet = "0123456789" + "ABCDEFGHIJKLMNOPQRSTUVWXYZ" + "abcdefghijklmnopqrstuvwxyz";
 
-    public static string EncodeToBase62(this int number)
+    public static string EncodeToBase62(this long number)
     {
         if (number == 0)
         {
@@ -14,7 +14,7 @@ public static class Base62EncodingExtensions
         var result = new Stack<char>();
         while (number > 0)
         {
-            result.Push(Alphabet[number % 62]);
+            result.Push(Alphabet[(int)number % 62]);
             number /= 62;
         }
